+++ src/lib/falcon_string.c
@@ -11,7 +11,7 @@
 /**
  * Hashes an input string using FNV-1a hash function.
  */
-uint32_t FalconHashString(const unsigned char *key, int length) {
+uint32_t FalconHashString(const unsigned char *key, size_t length) {
     uint32_t hash = 2166136261u;
 
     for (uint64_t i = 0; i < length; i++) {
@@ -26,7 +26,7 @@ uint32_t FalconHashString(const unsigned char *key, int length) {
  * Creates a new FalconObjString by claiming ownership of the given string. In this case, the
  * characters of a FalconObjString can be freed when no longer needed.
  */
-FalconObjString *FalconMakeString(FalconVM *vm, int length) {
+FalconObjString *FalconMakeString(FalconVM *vm, size_t length) {
     FalconObjString *str = (FalconObjString *) FalconAllocateObject(
         vm, sizeof(FalconObjString) + length + 1, FALCON_OBJ_STRING);
     str->length = length;
@@ -37,7 +37,7 @@ FalconObjString *FalconMakeString(FalconVM *vm, int length) {
  * Copies and allocates a given string to the heap. This way, every FalconObjString reliably owns
  * its character array and can free it.
  */
-FalconObjString *FalconCopyString(FalconVM *vm, const char *chars, int length) {
+FalconObjString *FalconCopyString(FalconVM *vm, const char *chars, size_t length) {
     uint32_t hash = FalconHashString((const unsigned char *) chars, length);
     FalconObjString *interned =
         FalconTableFindStr(&vm->strings, chars, length, hash); /* Checks if interned */
@@ -72,7 +72,7 @@ int FalconCompareStrings(const FalconObjString *str1, const FalconObjString *str
  */
 FalconObjString *FalconConcatStrings(FalconVM *vm, const FalconObjString *s1,
                                      const FalconObjString *s2) {
-    int length = s2->length + s1->length;
+    size_t length = s2->length + s1->length;
     FalconObjString *result = FalconMakeString(vm, length);
     memcpy(result->chars, s2->chars, s2->length);
     memcpy(result->chars + s2->length, s1->chars, s1->length);
